+++ lib/spout/tests/variable_type_validation.rb
@@ -1,7 +1,7 @@
 module Spout
   module Tests
     module VariableTypeValidation
-      VALID_VARIABLE_TYPES = ['identifier', 'choices', 'integer', 'numeric', 'string', 'text', 'date', 'time', 'file'].sort
+      VALID_VARIABLE_TYPES = ['identifier', 'choices', 'integer', 'numeric', 'string', 'text', 'date', 'time', 'file', 'datetime'].sort
 
       def assert_variable_type(item, msg = nil)
         full_message = build_message(msg, "? invalid variable type. Valid types: #{VALID_VARIABLE_TYPES.join(', ')}", item)