+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## 0.3.0
 
+### Enhancements
+- Use of Ruby 2.0.0-p247 is now recommended
+
 ## 0.2.0 (June 26, 2013)
 
 ### Enhancements