+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   - If a version is specified, `spout export 1.0.1` then the command line version is used
   - If no version is specified, and no `VERSION` file exists, then the default `1.0.0` is used
   - The version must be specified on the first line in the `VERSION` file
+- Spout can now create a new Spout project in a folder that was cloned from a Git repository
 - Use of Ruby 2.0.0-p247 is now recommended
 
 ## 0.2.0 (June 26, 2013)