+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Tests now include check for variable and domain name uniqueness across folders
   - `include Spout::Tests::VariableNameUniqueness`
   - `include Spout::Tests::DomainNameUniqueness`
+- Tests now allow `datetime` as a valid variable type
 - Exports will now create a folder based on the version specified in the `VERSION` file located in the root of the data dictionary
   - If a version is specified, `spout export 1.0.1` then the command line version is used
   - If no version is specified, and no `VERSION` file exists, then the default `1.0.0` is used